backend/requirements.txt
asgiref
Django
django-cors-headers #fixes cross origin request issue
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv